Move isTemporalAmount
